services/sfx-server/requirements.txt
# SFX server dependencies
# Note: EzAudio is installed in the Dockerfile from the pinned GitHub clone
# at /app/ezaudio (via pip install -r /app/ezaudio/requirements.txt).

# Server
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3

# Audio
torchaudio==2.5.1
soundfile==0.13.1

# HuggingFace model download (checkpoints auto-downloaded by EzAudio on first run)
huggingface_hub==0.30.2
hf_transfer==0.1.9
